Добавлены контроллеры, ДТО, мэпперы для сущностей отделение, банкомат, аудит

+++ public-info/src/main/java/com/bank/publicinfo/repository/BankDetailsRepository.java
@@ -3,5 +3,7 @@ package com.bank.publicinfo.repository;
 import com.bank.publicinfo.entity.BankDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BankDetailsRepository extends JpaRepository<BankDetails, Long> {
+import java.math.BigInteger;
+
+public interface BankDetailsRepository extends JpaRepository<BankDetails, BigInteger> {
 }
