+++ public-info/src/main/java/com/bank/publicinfo/repository/CertificateRepository.java
@@ -3,5 +3,7 @@ package com.bank.publicinfo.repository;
 import com.bank.publicinfo.entity.Certificate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CertificateRepository extends JpaRepository<Certificate, Long> {
+import java.math.BigInteger;
+
+public interface CertificateRepository extends JpaRepository<Certificate, BigInteger> {
 }
